Reapply "Deprecate spack solve" (#52596) (#52879)
This reverts commit 6e809d396521fd52ad221eb399ffa148b69251c7.
